# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BearLink API
  description: |
    REST API specification for BearLink URL shortening service.

    For async event documentation, see [AsyncAPI specification](./asyncapi.yaml).
  version: 2.0.0
  contact:
    name: BearLink Team
servers:
  - url: http://localhost:{port}
    description: Development server
    variables:
      port:
        enum:
          - '4000'
          - '5000'
          - '6000'
          - '7000'
        default: '5000'
...
paths:
...
  /urls:
    get:
      tags: [urls]
      summary: List user's URLs
      description: Returns paginated shortened URLs created by the authenticated user, with optional filtering.
      operationId: listUrls
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: tag
          in: query
          description: Filter by tag
          schema:
            type: string
            example: marketing
        - name: search
          in: query
          description: Search in originalUrl and customAlias (case-insensitive)
          schema:
            type: string
            example: example.com
        - name: expired
          in: query
          description: Filter by expiry status
          schema:
            type: string
            enum: ['true', 'false']
            example: 'false'
      responses:
        '200':
          description: Paginated list of URLs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUrlsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [urls]
      summary: Create shortened URL
      description: |
        Creates a new shortened URL for the authenticated user.

        The destination URL is validated against the Google Safe Browsing API (if configured)
        and domain allowlist/blocklist rules.

        If `requireSignature` is set to `true`, the response also includes a pre-signed
        `signedUrl` that embeds an expiring HMAC signature. Unsigned requests to this link
        will be rejected with 403.
      operationId: createUrl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUrlRequest'
      responses:
        '200':
          description: URL created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUrlResponse'
        '400':
          description: Validation error (missing fields, invalid format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Custom alias already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: URL rejected by domain filter or Safe Browsing check
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /urls/bulk:
    post:
      tags: [urls]
      summary: Bulk create shortened URLs
      description: |
        Creates up to 50 shortened URLs in a single request.
        Each item is processed independently — if one fails validation or DB insertion,
        the others still succeed. Returns a `results` array where each entry is either
        `{ shortUrl }` on success or `{ error }` on failure.
      operationId: createUrlsBulk
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCreateUrlRequest'
      responses:
        '200':
          description: Partial or full success — check individual result entries for errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkCreateUrlResponse'
        '400':
          description: Top-level validation error (not an array, empty, or exceeds 50)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /urls/{id}:
    put:
      tags: [urls]
      summary: Update URL
      description: |
        Updates an existing shortened URL. Only the owner can update their URL.

        The new destination URL is validated against the domain filter and Safe Browsing API.
        Cache entries for this URL are invalidated on success.
      operationId: updateUrl
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/urlId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUrlRequest'
      responses:
        '200':
          description: URL updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Url'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: URL not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Custom alias already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: URL rejected by domain filter or Safe Browsing check
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [urls]
      summary: Delete URL
      description: Permanently deletes a shortened URL. Cache entries are invalidated on success.
      operationId: deleteUrl
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/urlId'
      responses:
        '204':
          description: URL deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: URL not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /urls/{id}/sign:
    post:
      tags: [urls]
      summary: Generate a signed URL
      description: |
        Creates a time-limited, HMAC-signed version of the short URL.
        The signed URL includes `?exp=<unix-timestamp>&sig=<hmac-sha256>` query parameters.

        Requires `URL_SIGNING_SECRET` to be configured on the server.
        Signed URLs are useful for distributing links that expire automatically
        or that you want to prevent from being shared beyond the intended recipient.
      operationId: signUrl
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/urlId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUrlRequest'
      responses:
        '200':
          description: Signed URL generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUrlResponse'
        '400':
          description: Invalid URL ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: URL not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: URL signing not configured (missing URL_SIGNING_SECRET)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Public Redirect Endpoints ───────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /login or /register
  parameters:
...
    urlId:
      name: id
      in: path
      required: true
      description: URL record ID
      schema:
        type: integer
        example: 1
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    CreateUrlRequest:
      type: object
      required:
        - originalUrl
      properties:
        originalUrl:
          type: string
          format: uri
          description: The URL to shorten (must be http or https)
          example: https://example.com/very/long/path
        redirectType:
          type: integer
          enum: [301, 302]
          default: 302
          description: HTTP redirect type (301 = permanent, 302 = temporary)
          example: 302
        customAlias:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]{3,50}$'
          description: Custom slug to use instead of a generated short ID
          example: my-campaign
        expiresAt:
          type: string
          format: date-time
          description: When the link expires (must be in the future)
          example: '2027-01-01T00:00:00.000Z'
        password:
          type: string
          description: Protect the link with a password (stored as bcrypt hash)
          example: secret123
        tags:
          type: array
          items:
            type: string
          description: Arbitrary string tags for filtering and organisation
          example: [marketing, q1]
        utmParams:
          type: object
          description: |
            UTM parameters to append to the destination URL on redirect.
            All values must be strings or numbers.
          additionalProperties:
            type: string
          example:
            utm_source: newsletter
            utm_medium: email
            utm_campaign: spring-sale
        requireSignature:
          type: boolean
          default: false
          description: |
            If true, redirect requests must include a valid HMAC signature
            (obtained from POST /urls/:id/sign or returned in the creation response).
          example: false
        signedTtl:
          type: integer
          minimum: 1
          description: |
            Validity window in seconds for the initial signed URL returned in the response
            (only relevant when requireSignature=true). Defaults to 3600 (1 hour).
          example: 86400
    UpdateUrlRequest:
      type: object
      required:
        - originalUrl
      properties:
        originalUrl:
          type: string
          format: uri
          description: New destination URL
          example: https://example.com/updated/path
        redirectType:
          type: integer
          enum: [301, 302]
          description: HTTP redirect type
          example: 302
        customAlias:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]{3,50}$'
          nullable: true
          description: New custom alias, or null to remove the existing alias
          example: new-alias
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: New expiry (must be in the future), or null to remove expiry
          example: '2027-06-01T00:00:00.000Z'
        password:
          type: string
          nullable: true
          description: New password, or null to remove password protection
          example: newpassword
        tags:
          type: array
          items:
            type: string
          description: Replace all tags with this list
          example: [social, organic]
        utmParams:
          type: object
          nullable: true
          description: Replace UTM params with this object, or null to remove them
          additionalProperties:
            type: string
          example:
            utm_source: twitter
        requireSignature:
          type: boolean
          description: Enable or disable signature enforcement
          example: true
    BulkCreateUrlRequest:
      type: object
      required:
        - urls
      properties:
        urls:
          type: array
          minItems: 1
          maxItems: 50
          description: Array of URL creation requests (max 50)
          items:
            $ref: '#/components/schemas/CreateUrlRequest'
    BulkCreateUrlResponse:
      type: object
      properties:
        results:
          type: array
          description: |
            One entry per input item. Each entry is either a success
            `{ shortUrl }` or a failure `{ error }`.
          items:
            oneOf:
              - type: object
                properties:
                  shortUrl:
                    type: string
                    format: uri
                    example: http://localhost:5000/abc123xyz9
              - type: object
                properties:
                  error:
                    type: string
                    example: Custom alias is already taken.
    CreateUrlResponse:
      type: object
      properties:
        shortUrl:
          type: string
          format: uri
          description: The full shortened URL
          example: http://localhost:5000/abc123xyz9
        signedUrl:
          type: string
          format: uri
          description: |
            Pre-signed URL with HMAC signature appended (only present when
            requireSignature=true and URL_SIGNING_SECRET is configured).
          example: http://localhost:5000/abc123xyz9?exp=1740700000&sig=a3f9c2...
    SignUrlRequest:
      type: object
      properties:
        ttl:
          type: integer
          minimum: 1
          description: Validity window in seconds (default 3600)
          example: 86400
    SignUrlResponse:
      type: object
      properties:
        signedUrl:
          type: string
          format: uri
          description: The signed URL with ?exp and ?sig query parameters appended
          example: http://localhost:5000/abc123xyz9?exp=1740786400&sig=7b2e41...
...
    Url:
      type: object
      description: A shortened URL record
      properties:
        id:
          type: integer
          example: 1
        originalUrl:
          type: string
          format: uri
          example: https://example.com/very/long/path
        shortId:
          type: string
          description: Auto-generated 10-character identifier
          example: abc123xyz9
        customAlias:
          type: string
          nullable: true
          description: Custom slug (3-50 chars, alphanumeric + hyphen + underscore)
          example: my-campaign
        redirectType:
          type: integer
          enum: [301, 302]
          example: 302
        expiresAt:
          type: string
          format: date-time
          nullable: true
          example: '2027-01-01T00:00:00.000Z'
        tags:
          type: array
          items:
            type: string
          example: [marketing, q1]
        utmParams:
          type: object
          nullable: true
          additionalProperties:
            type: string
          example:
            utm_source: newsletter
            utm_medium: email
        requireSignature:
          type: boolean
          default: false
          example: false
        clicks:
          type: integer
          minimum: 0
          example: 42
        userId:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000Z'
        previewTitle:
          type: string
          nullable: true
          description: Page title scraped by preview-service (async)
          example: Example Domain
        previewDescription:
          type: string
          nullable: true
          description: Meta description scraped by preview-service (async)
          example: This domain is for use in illustrative examples.
        previewImageUrl:
          type: string
          nullable: true
          description: OG image URL scraped by preview-service (async)
          example: https://example.com/og-image.png
        previewFetchedAt:
          type: string
          format: date-time
          nullable: true
          description: When the preview metadata was last fetched
          example: '2024-01-15T10:31:00.000Z'
    ListUrlsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Url'
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              example: 87
            pages:
              type: integer
              example: 5

    # ── Analytics ─────────────────────────────────────────────────────────────
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Invalid credentials